% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{calculate_data_grid}
\alias{calculate_data_grid}
\title{Add transformation values to each observation in data_grid}
\usage{
calculate_data_grid(grid)
}
\arguments{
\item{grid}{An SVFGrid object.}
}
\value{
The modified grid object with phi and c_cells results added.
}
\description{
This function processes each observation in \code{data_grid} using the specified
inputs, calculating phi values and contiguous cells, and updates the grid object.
}
\examples{

data <- data.frame(x = c(1, 3, 5, 7), y = c(2, 4, 6, 8))
inputs <- c("x")
outputs <- c("y")
d <- 2
grid <- SVFGrid(data, inputs, outputs, d)
grid <- create_grid(grid)
grid <- calculate_data_grid(grid)

}
