% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{calculate_dmu_phi}
\alias{calculate_dmu_phi}
\title{Calculate the transformation (phi) value of an observation in the grid}
\usage{
calculate_dmu_phi(grid, cell)
}
\arguments{
\item{grid}{An SVFGrid object.}

\item{cell}{Position of the observation in the grid.}
}
\value{
A list containing the phi values for the specified cell.
}
\description{
This function computes and returns the phi value for a specific grid cell,
based on the grid data and the cell position.
}
\examples{

data <- data.frame(x = c(1, 3, 5), y = c(2, 4, 6))
inputs <- c("x")
outputs <- c("y")
d <- 2
grid <- SVFGrid(data, inputs, outputs, d)
grid <- create_grid(grid)
cell <- c(1)
phi_result <- calculate_dmu_phi(grid, cell)
print(phi_result)

}
