% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{calculate_grid_properties}
\alias{calculate_grid_properties}
\title{Add transformation values (phi) to the grid dataframe}
\usage{
calculate_grid_properties(grid)
}
\arguments{
\item{grid}{An SVFGrid object.}
}
\value{
The updated SVFGrid object with the grid dataframe.
}
\description{
This function calculates and adds additional information to the grid's
dataframe, including the phi values for each cell and the contiguous cells.
}
\examples{

data <- data.frame(x = c(1, 3, 5), y = c(2, 4, 6))
inputs <- c("x")
outputs <- c("y")
d <- 2
grid <- SVFGrid(data, inputs, outputs, d)
grid <- create_grid(grid)
grid <- calculate_grid_properties(grid)

}
