% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{create_grid_svfgrid}
\alias{create_grid_svfgrid}
\title{Create a grid based on data and hyperparameter d}
\usage{
create_grid_svfgrid(grid)
}
\arguments{
\item{grid}{An SVFGrid object.}
}
\value{
The SVFGrid object with the created grid.
}
\description{
This function creates a grid using the provided data and hyperparameter \code{d}.
The grid represents the data in a space divided into cells defined by \code{d}.
}
\examples{

data <- data.frame(x1 = c(1, 2, 3, 4), x2 = c(2, 4, 3, 2), y1 = c(1, 1, 2, 3))

inputs <- c("x1", "x2")
outputs <- c("y1")
d <- 2
grid_obj <- SVFGrid(data, inputs, outputs, d)
grid_obj <- create_grid_svfgrid(grid_obj)

}
