% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf.R
\name{create_matrix_partitions}
\alias{create_matrix_partitions}
\title{Title create_matrix_partitions}
\usage{
create_matrix_partitions(x, d)
}
\arguments{
\item{d:}{}
}
\value{

}
\description{
Función que crea las particiones de cada dimensión. Coge cada dimensión de inputs y la trocea en d particiones equidistantes
}
\examples{

}
