% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf.R
\name{get_estimation}
\alias{get_estimation}
\title{Estimation of a DMU}
\usage{
get_estimation(svf, dmu)
}
\arguments{
\item{svf}{SVF object.}

\item{dmu}{A numeric vector with the characteristics of the DMU (Must have the same number of elements as the inputs).}
}
\value{
A numeric vector with the estimations for each output.
}
\description{
This function calculates the output estimations for a specific DMU using the SVF model.
}
