% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{GRID}
\alias{GRID}
\title{GRID Class for SVF model}
\usage{
GRID(data, inputs, outputs, d)
}
\arguments{
\item{data}{DataFrame with the dataset on which the grid is built.}

\item{inputs}{List of inputs.}

\item{outputs}{List of outputs.}

\item{d}{Number of partitions in which the grid is divided.}
}
\description{
This class represents a grid over which the SVF model is applied.
A grid is a partition of the input space divided into cells.
}
\details{
To create a GRID object, provide a DataFrame with the data,
the names of the input (inputs) and output (outputs) variables, and
the number of partitions (d).
}
\section{Fields}{

\describe{
\item{\code{data_grid}}{Data grid, initially NULL.}

\item{\code{knot_list}}{List of grid nodes, initially NULL.}
}}

\examples{
data <- data.frame(x = sample(1:10, 4, replace = TRUE),
                   y = sample(1:10, 4, replace = TRUE))
grid <- GRID(data, inputs = c("x"), outputs = c("y"), d = 2)
print(grid)

}
