% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvf.R
\name{plot_svf_estimation}
\alias{plot_svf_estimation}
\title{Plot Support Vector Frontier (SVF) Estimation}
\usage{
plot_svf_estimation(svf, input1, input2 = NULL, grid_size = 10)
}
\arguments{
\item{svf}{A list object containing the solution to a Support Vector Frontier model. It must contain a resolved model, which can be obtained by calling the \code{solve} function. This object should have a \code{solution} field, input data (\code{data}), and output data (\code{outputs}).}

\item{input1}{A string representing the name of the first input variable in the \code{svf$data} list. This variable will be plotted on the x-axis of the plot.}

\item{input2}{A string representing the name of the second input variable in the \code{svf$data} list. This is only required if you want to generate a 2D plot. If not provided, the function will generate a 1D plot. Default is \code{NULL}.}

\item{grid_size}{A numeric value that defines the resolution of the grid for the estimation. It determines how many points are used to generate the estimation. The default value is 10.}
}
\value{
A Plotly object representing the generated plot. The plot includes:
\itemize{
\item A line representing the estimated SVF (in 1D) or a surface representing the estimated SVF (in 2D).
\item Markers representing the Data Management Units (DMUs) from the \code{svf$data}.
}
}
\description{
This function generates a plot of the estimated Support Vector Frontier (SVF) for one or two input variables.
The plot is either 1D or 2D depending on whether one or two input variables are provided.
The estimation is calculated using the \code{get_estimation_svf} function for each point in the input variable range(s).
}
\details{
If only one input variable (\code{input1}) is specified, the function will create a 1D plot of the SVF estimation across the range of the \code{input1} variable.
If both \code{input1} and \code{input2} are specified, the function will generate a 2D surface plot of the SVF estimation over a grid defined by the \code{input1} and \code{input2} ranges.
The \code{get_estimation_svf} function is called to estimate the SVF for each point in the specified grid.
}
\examples{
data <- data.frame(
x1 = c(1, 2, 3, 10),
x2 = c(2, 4, 3, 10),
y1 = c(1, 1, 2, 3))
method <- "ssvf"
inputs <- c("x1", "x2")
outputs <- c("y1")
c_value <- 1
eps_value <- 0.1
d_value <- 2

svf_model <- SSVF(method, inputs, outputs, data, c_value, eps_value, d_value)

svf_model <- train(svf_model)


svf_model <- solve(svf_model)

fig <- plot_svf_estimation(svf_model, "x1", "x2",  grid_size = 100)

fig

}
