% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf.R
\name{print_ssvf_model}
\alias{print_ssvf_model}
\title{Prints the optimization model in a human-readable format}
\usage{
print_ssvf_model(svf, bounds = NULL)
}
\arguments{
\item{svf}{SSVF object with the trained model.}

\item{bounds}{A list of bounds for each variable, e.g., list(c(0, 10), c(1, 5)).}
}
\description{
This function prints the optimization problem of the SSVF model in a format similar to the one you provided.
It is fully modular, meaning it works for any SSVF optimization problem.
}
