% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{search_contiguous_cell}
\alias{search_contiguous_cell}
\title{Find contiguous cells in SVFGrid}
\usage{
search_contiguous_cell(cell)
}
\arguments{
\item{cell}{Vector specifying the cell position in the grid.}
}
\value{
A list of cells contiguous to the specified cell.
}
\description{
This function identifies and returns the cells contiguous to a specified cell
in the grid. Contiguous cells share at least one edge or vertex with the cell.
}
\examples{

cell <- c(2, 2)
contiguous_cells <- search_contiguous_cell(cell)

}
