% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{search_dmu}
\alias{search_dmu}
\title{Function to return the cell in which an observation is located in the grid}
\usage{
search_dmu(grid, dmu)
}
\arguments{
\item{grid}{GRID object.}

\item{dmu}{Observation to search for in the grid.}
}
\value{
Vector with the position of the observation in the grid.
}
\description{
This method searches for a specific DMU in the grid and returns the cell
in which that observation is located.
}
\examples{
data <- data.frame(x1 = c(1, 2, 3, 4),
                  x2 = c(2, 4, 3, 2),
                  x3 = c(1, 1, 2, 3),
                  y1 = c(1, 1, 2, 3))

inputs <- c("x1", "x2")
outputs <- c("y1")
d <- 2

grid_instance <- GRID(data, inputs, outputs, d)

grid_instance$knot_list <- list(list(1, 2.5, 4), list(2, 3, 4), list(1,2,3))
dmu <- c(3, 5, 1)
position <- search_dmu(grid_instance, dmu)

print(paste("Position in the grid: (", paste(position, collapse = ", "), ")", sep = ""))

}
