% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvf.R
\name{solve}
\alias{solve}
\title{Solves the SSVF model}
\usage{
solve(svf)
}
\arguments{
\item{svf}{SSVF object with a trained model.}
}
\value{
A list containing the solutions for the 'w' and 'xi' variables.
}
\description{
This function solves the previously trained SSVF model.
It extracts the solution from the model after it has been trained.
}
