% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf.R
\name{SVF}
\alias{SVF}
\title{Creates a new instance of the SVF class}
\usage{
SVF(method, inputs, outputs, data, C, eps, d)
}
\arguments{
\item{method}{SVF method to be used.}

\item{inputs}{Inputs to evaluate in the dataset.}

\item{outputs}{Outputs to evaluate in the dataset.}

\item{data}{Dataset to evaluate.}

\item{C}{Values of the model's hyperparameter C.}

\item{eps}{Values of the model's epsilon hyperparameter.}

\item{d}{Value of the model's hyperparameter d.}
}
\value{
An object of class 'SVF'.
}
\description{
This function defines the base SVF model, with common properties and functions.
It acts as an abstract class and should not be instantiated directly.
}
