% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf_primal_solution.R
\name{SVFPrimalSolution}
\alias{SVFPrimalSolution}
\title{Creates an SVFPrimalSolution object}
\usage{
SVFPrimalSolution(w, xi)
}
\arguments{
\item{w}{Values of the weights w from the solved model.}

\item{xi}{Values of xi from the solved model.}
}
\value{
An object of class SVFPrimalSolution.
}
\description{
Creates an SVFPrimalSolution object
}
\examples{

w_values <- c(0.5, 1.5, 2.5)
xi_values <- c(0.1, 0.2, 0.3)
solution <- SVFPrimalSolution(w_values, xi_values)
class(solution)
solution$w
solution$xi

}
