% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvf.R
\name{train}
\alias{train}
\title{Trains the SSVF model}
\usage{
train(svf)
}
\arguments{
\item{svf}{SSVF object.}
}
\value{
The SSVF object with the model configuration (without solution).
}
\description{
This function generates the SSVF model but does not solve the optimization problem.
It only prepares the model to be solved later.
}
