% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{GRID}
\alias{GRID}
\title{Clase GRID para modelo SVF}
\usage{
GRID(data, inputs, outputs, d)
}
\arguments{
\item{data}{DataFrame con el conjunto de datos sobre los que se construye el grid.}

\item{inputs}{Listado de inputs.}

\item{outputs}{Listado de outputs.}

\item{d}{número de particiones en las que se divide el grid.}
}
\description{
Esta clase representa un grid sobre el que se realiza el modelo SVF.
Un grid es una partición del espacio de los inputs dividido por celdas.
}
\details{
Para crear un objeto GRID, proporcione un DataFrame con los datos,
los nombres de las variables de entrada (inputs) y salida (outputs), y
el número de particiones (d).
}
\section{Fields}{

\describe{
\item{\code{data_grid}}{Grid de datos, inicialmente NULL.}

\item{\code{knot_list}}{Lista de nodos del grid, inicialmente NULL.}
}}

\examples{
data <- data.frame(x = rnorm(10), y = rnorm(10))
grid <- GRID(data, inputs = c("x"), outputs = c("y"), d = 4)
print(grid)

data <- data.frame(x = rnorm(10), y = rnorm(10))
grid <- GRID(data, inputs = c("x"), outputs = c("y"), d = 4)
print(grid)
}
