% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvf.R
\name{SSVF}
\alias{SSVF}
\title{Creates an SSVF object}
\usage{
SSVF(method, inputs, outputs, data, c, eps, d)
}
\arguments{
\item{method}{SVF method to be used.}

\item{inputs}{Inputs to evaluate in the dataset.}

\item{outputs}{Outputs to evaluate in the dataset.}

\item{data}{Dataset to evaluate.}

\item{c}{Values of the model's hyperparameter C.}

\item{eps}{Values of the model's epsilon hyperparameter.}

\item{d}{Value of the model's hyperparameter d.}
}
\value{
An object of class SSVF.
}
\description{
Creates an SSVF object
}
\examples{

data <- data.frame(
  x1 = c(1, 2, 3, 10),
  x2 = c(2, 4, 3, 10),
  y1 = c(1, 1, 2, 3)
)

method <- "ssvf"
inputs <- c("x1", "x2")
outputs <- c("y1")
c_value <- 2.0
eps_value <- 0
d_value <- 4

svf_model <- SSVF(method, inputs, outputs, data, c_value, eps_value, d_value)

svf_model <- train(svf_model)

print_model(svf_model)

svf_model <- solve(svf_model)

print(svf_model$solution$w)

print(svf_model$solution$xi)

}
