% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf.R
\name{transformation}
\alias{transformation}
\title{Title transformation}
\usage{
transformation(x_j, t_l_j)
}
\arguments{
\item{x_j:}{valor de la celda a evaluar}

\item{t_l_j:}{valor del nodo con el que se quiere comparar}
}
\value{
res: resultado de la transformación
}
\description{
Esta función evalúa si el valor de una celda es mayor o menor al de un nodo del grid.
Si es mayor devuelve 1, si es igual devuelve 0 y si es menor devuelve -1.
}
\examples{

}
