% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{transformation}
\alias{transformation}
\title{Transformation of values in the GRID}
\usage{
transformation(x_i, t_k)
}
\arguments{
\item{x_i}{Value of the cell to evaluate.}

\item{t_k}{Value of the node to compare with.}
}
\value{
Result of the comparison: 1, 0, -1.
}
\description{
This function evaluates whether the value of an observation is greater than, equal to,
or less than the value of a node in the grid. It returns 1 if greater,
0 if equal, and -1 if less.
}
\examples{
result <- transformation(2, 5)
cat("Transformation result:", result, "\n")

result <- transformation(3, 3)
cat("Transformation result:", result, "\n")

result <- transformation(4, 2)
cat("Transformation result:", result, "\n")

}
